+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Generate warning when user attempts a Box-Cox transformation of non-positive data (@LiamBlake, #713).
 
+* Added a new step called `step_dummy_multi_choice()`, which will take multiple nominal variables and produces shared dummy variables. (#716)
+
 # recipes 0.1.16
 
 ## New Steps